+++ .vscode-test.mjs
@@ -3,7 +3,7 @@ import { defineConfig } from '@vscode/test-cli';
 export default defineConfig({
   files: 'out/test/**/*.test.js',
   mocha: {
-    timeout: 1000000,
+    timeout: 1000000
   },
+  launchArgs: ['--disable-extensions']
 });
-