+++ src/test/components/destructureNested/destructureNested.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Component() {
+  const {
+    styles: { color },
+    text
+  } = { styles: { color: 'red' }, text: 'Test' };
+
+  return <div className={color}>{text}</div>;
+}
+