+++ src/test/components/destructureRename/destructureRename.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function Component() {
+  const { colour: color } = { colour: 'red' };
+
+  return <div className={color}>Test</div>;
+}
+